Header and Hero design
Header and Hero has been placed where the logo and the links are at the
extremes of the header and the hero contains two sections, one on the
left with information and one on the right as a placeholder for an
image, these two take all the space available in the width.
Button has been styled.
space between elements in hero-left have been styled.